+++ lib/rdmm/responses/list_makers_response.rb
@@ -1,11 +1,16 @@
 module Rdmm
   module Responses
     class ListMakersResponse < BaseResponse
+      private
+
       # @note Override
-      def resources
-        body["result"]["maker"].map do |source|
-          ::Rdmm::Resources::MakerResource.new(source)
-        end
+      def resource_class
+        ::Rdmm::Resources::MakerResource
+      end
+
+      # @note Override
+      def sources
+        body["result"]["maker"]
       end
     end
   end