+++ lib/rdmm/responses/list_series_response.rb
@@ -1,11 +1,16 @@
 module Rdmm
   module Responses
     class ListSeriesResponse < BaseResponse
+      private
+
       # @note Override
-      def resources
-        body["result"]["series"].map do |source|
-          ::Rdmm::Resources::SeriesResource.new(source)
-        end
+      def resource_class
+        ::Rdmm::Resources::SeriesResource
+      end
+
+      # @note Override
+      def sources
+        body["result"]["series"]
       end
     end
   end