+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.2.0
 
+- Add some methods to `Rdmm::Responses::BaseResponse` for error handling
 - Rename `:floor` to `:floor_code` on `Rdmm::Client#list_items`
 - Rename `:service` to `:service_code` on `Rdmm::Client#list_items`